requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
dotenv
apscheduler
openai
python-multipart # For form data/uploads
email-validator # For email validation in pydantic/sqlmodel


pydantic-settings


psutil

requests

jinja2
